Simplify affixes and combat mechanics.
Instead of AffixTypes which are generators for affixes, there is just a
fixed set of affixes, each with a fixed set of stats. "of Maiming"
always has the exact same effect.

If we want to make variants of an affix that look the same but have
different numbers, we can just make multiple affixes with the same name
but different values. (We'll have to also give them a hidden unique
name for storage and storing in the ResourceSet.)

Also, make some of the combat numbers ints. Trying to simplify things
down so there are fewer numbers to tune, and an easier set of modifiers
to display to users.