Change math and optimize random resource choice.
The heart of room, item, affix, and monster generation is ResourceSet,
and its methods for choosing a random object at a given depth. Its goal
is to prefer selecting an object near that depth, with some sort of
bell curve-like fall off on either side. It used an accumulation of
weird hacks and goofy math to approximate that. It was also slow.

Now, each resource has an actual normal distribution centered on its
level. The left and right sides of the bell curve are carefully tweaked
to, I hope give nice odds of choosing out of depth (above and below)
stuff, taking into account the depth the resource is being chosen for.

Doing this "properly" involves a good bit of math and creating a
probability table of all of the resources. That's slow. Fortunately, we
can cache that and reuse it across calls.

With that caching in place, and a binary search over the weighted
table, this now selects resources about 20x faster than the old code.