Split out Attack into separate Attack and Hit classes.
The former is the immutable fundamental damaging properties of an item.
The latter is a mutable object built from an attack with a number of
modifiers applied to it. It represents a single use of an attack in a
certain context.

Trying to use one class for both was always a little weird. When I
started trying to support things like affixed armor that improved
damage, it got even weirder.

This cleans it up, I think. It also enables things like equipment
bonuses affecting ranged attacks and using magical items.