Move monsters out of encounters.
Instead, move minions, stains, and location preferences into Breed.
It seems to make sense there since even the multi-monster encounters
still had a single "lead" monster, and is much simpler.

"Encounters" is now simplified to just the items that are dropped on
the floor, currently corpses and rocks. My expectation is that this
will be used mainly to tune the essential items needed during
exploration: food, light, etc.

In addition, when a dungeon is generated, some fraction of the monsters
are immediately slain, leaving their stain and drops, if any. That
peppers some more stuff on the floor, leaves empty stains, while still
choosing items that are appropriate to the depth.