Take strength and heft into account with weapon damage.
Higher strength increases damage, but each weapon also has a "heft"
value that is the zero point for how much strength the hero must have
to do 1.0 damage.

This, along with encumbrance means that there is tension between
strength, weapons, and armor.

- Heavier armor increases protection, but lowers strength, which in
  turn lowers damage.
- More powerful weapons have higher heft, which require greater strength
  to use effectively.
- Increasing strength eases both of these but takes points that can't
  be spent on other attributes.

All of this helps the different weapons have different trade-offs. The
ideal weapon for a hero depends heavily on their strength and armor and
no weapon category is dramatically better than others.