Motilities!
Instead of an increasingly long list of flags like "canOpenDoors" and
"canFly" passed to pathfinding, flow, etc. there is now the concept of
"motility". Each motility is a way an actor may be able to move, like
"walk", "fly", "openDoor", "swim".

Each actor has a set of motilities they support. Each tile has a set
of motilities which can be used to enter it. Pathfinding and other
traversals check that the actor has at least one of the motilities that
the tile requires.

This should make it easier to add other terrains and special monsters
like lava pools and fire elementals, etc.

In short, frogs can now spawn in water and swim.