Redo room themes.
Instead of painting them on after the fact, pick a theme when initially
placing the room. This is simpler, more powerful, and handles the
relationships between rooms better.

It will let us define custom classes for rooms that should be more
interesting then just some decor. Things like:

- Great halls with symmetric tables down the center.
- Halls with symmetric doors.
- Special shaped large rooms like conservatories, jails, etc.

It also uses ResourceSet to let us tune the frequency of room types and
have it change based on depth.