Revamp the monster debugging UI.
The new vision and hearing AI I added a while back is sorely in need of
tweaking and tuning. I need to tune the different breeds so that they
don't all have max hearing and vision.

But I also need to make sure the core AI that uses those tuned values
works well. My impression is that it's off balance. It seems like
monsters are almost always waiting at the door when the player opens it.

Part of the problem is that it's hard to see what the AI is doing. I had
some old debug code to show that but it bit rotted. So this revamps it
and gets it in a good state again.

When Debug.enabled is set, you can click a monster to turn on its debug
pane. That shows a little bar chart of various AI values over time and
the reasons for their most recent update. It took a lot of iteration
before I got to something that felt like it showed the right amount of
detail. It's easy for it to be not enough, or to clutter the screen to
unreadability.

I also added a wizard mode flag to toggle showing all monsters
regardless of visibility. That lets me debug monsters on the other side
of doors. It's really monsters the hero *can't* see where alertness
comes into play.

As soon as I got this working, I could immediately tell there is a
problem in the alertness. If a monster ever wakes up, the alertness
decay is too low and it basically stays awake forever unless the hero
gets really far away, even if the hero is resting.

So the next step is actually using these new tools to try to fix and
tune all of this.