Solidify the code around placing unstyle tiles.
The temporary tile types now live in a separate class and all unpainted
tiles are temporary types so there is a clear distinction between
unpainted and styled.

Added a "wet passage" type to handle unstyled bridges.

Tweaked the ownership code so that tiles are always owned by an
architecture when appropriate, including aquatic tiles. That should make
them easier to style.